requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.3.0
langchain-community==0.3.0
redis==5.0.1
python-dotenv==1.0.0
requests==2.31.0
asyncio-mqtt==0.16.1
websockets==12.0
aiofiles==23.2.1
dashscope==1.17.0
pydantic==2.5.0